✨ feat: Adicionando o loading para melhorar a experiência do usuário
